+++ src/pdfPreview.ts
@@ -462,6 +462,7 @@ export class PdfPreview extends Disposable {
 <script src="${resolveAsUri('lib', 'highlight.js')}"></script>
 <script src="${resolveAsUri('lib', 'toolbar.js')}"></script>
 <script src="${resolveAsUri('lib', 'undo.js')}"></script>
+<script src="${resolveAsUri('lib', 'noPrint.js')}"></script>
 </head>`;
 
     const body = `<body tabindex="1">